Refactor use of variable names in action
Moved again the action edit files to make the heirarchy clearer.
Now use the typehelper property info class to pass variable names to condition UC editor
That causes a lot of churn due to it being passed down in the configurationmenu.

The ActionPack UCs have a onAdditionaNames which ask the Components for list of names.  These are assigned to the ConditionFilterUC.VariableNames on use of that form
EventProgramCond & EventVoice supply that call back and use the event name to pass back up to actioncontroller to supply the variables.